+++ @navikt/core/react/src/overlays/dismissablelayer/DismissableLayer.stories.tsx
