+++ @navikt/core/react/src/overlays/dismissablelayer/DismissableLayer.tsx
