+++ @navikt/core/react/src/overlays/dismissablelayer/README.md
