+++ @navikt/core/react/src/overlays/dismissablelayer/util/dispatchCustomEvent.ts
