+++ @navikt/core/react/src/overlays/dismissablelayer/util/useEscapeKeydown.ts
