+++ @navikt/core/react/src/overlays/dismissablelayer/util/useFocusOutside.ts
