+++ @navikt/core/react/src/overlays/dismissablelayer/util/usePointerDownOutside.ts
